[Refactory] 싱글톤 상속 제거, 각 클래스별 세팅
