--- Assets/02. Scripts/Dohyeun/Common/LDHSingletonBehavior.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 8f3472641eaeb1b408e9061d0a3589b8
