+++ Assets/02. Scripts/Dohyeun/Managers/AudioManager.cs
@@ -14,8 +14,51 @@ public enum SFX // 네이밍 = 파일명
     sfx_test_click,
     COUNT,
 }
-public class AudioManager : LDHSingletonBehavior<AudioManager>
+public class AudioManager : MonoBehaviour
 {
+
+    #region 싱글톤 공통
+    public static AudioManager instance;
+
+    private bool isDestroyOnLoad = false;
+    private bool isInitialized = false;
+
+    private void Awake()
+    {
+        if (instance == null)
+        {
+            instance = this;
+            if (!isDestroyOnLoad)
+            {
+                if (transform.parent)
+                    DontDestroyOnLoad(transform.parent);
+                else
+                    DontDestroyOnLoad(this);
+            }
+        }
+        else
+        {
+            if (instance != this)
+            {
+                Destroy(gameObject);
+            }
+        }
+        if (!isInitialized)
+            instance.Initialize();
+        isInitialized = true;
+    }
+    #endregion
+    private void Initialize()
+    {
+        // 게임 시작 시, 데이터 불러오기
+        LoadBGMPlayer();
+        LoadSFXPlayer();
+        foreach (AudioSource audioSource in bgmPlayer.Values)
+            audioSource.volume = 0f;
+        foreach (AudioSource audioSource in sfxPlayer.Values)
+            audioSource.volume = masterVolume;
+    }
+
     // Resources 내 데이터 경로
     private const string AUDIO_PATH = "DohyeunTest/Audio";
 
@@ -31,17 +74,6 @@ public class AudioManager : LDHSingletonBehavior<AudioManager>
     public float masterVolume = 0.8f;
     //public float bgmMaxVolume = 1f;
     //public float sfxMaxVolume = 1f;
-
-    protected override void InitChild()
-    {
-        // 게임 시작 시, 데이터 불러오기
-        LoadBGMPlayer();
-        LoadSFXPlayer();
-        foreach (AudioSource audioSource in bgmPlayer.Values)
-            audioSource.volume = 0f;
-        foreach (AudioSource audioSource in sfxPlayer.Values)
-            audioSource.volume = masterVolume;
-    }
 
     private void LoadBGMPlayer()
     {
